requirements.txt
streamlit
google-generativeai
transformers
requests
